feat(midnight-sdk): address review feedback
