+++ packages/midnight-sdk/src/fetch/Ratifier.ts
@@ -6,10 +6,18 @@ import {
   RatifierUtils,
 } from "../signatures/RatifierUtils.js";
 import type { MidnightCallParameters } from "./types.js";
-import { bytecodeCallParameters, resolveChainId } from "./utils.js";
+import { resolveChainId } from "./utils.js";
 
 /**
- * Fetches maker bytecode and classifies the ratifier route.
+ * Fetches maker bytecode and classifies the ratifier route for new offers.
+ *
+ * The `bytecode` read is the result of `eth_getCode` for `params.maker` at the
+ * requested block. It is not returned to callers; it only tells the SDK whether
+ * the maker is an EOA or EIP-7702 account that can use Ecrecover signatures, or
+ * a deployed-code account that should use Setter root approval. Put the returned
+ * `ratifier` address on `Offer.create`, then use the returned `type` to choose
+ * `EcrecoverRatifierUtils.ratify` or `SetterRatifierUtils.ratify` after the tree
+ * has been built.
  *
  * @param client - Viem client used for the bytecode read.
  * @param params - Fetch parameters.
@@ -20,8 +28,11 @@ import { bytecodeCallParameters, resolveChainId } from "./utils.js";
  * ```ts
  * import { fetchRatifierInfo } from "@morpho-org/midnight-sdk";
  *
- * const info = await fetchRatifierInfo({} as never, {} as never);
- * console.log(info.type);
+ * const maker = "0x7b093658BE7f90B63D7c359e8f408e503c2D9401";
+ * const info = await fetchRatifierInfo(client, { maker });
+ * // Pass info.ratifier to Offer.create(...), then use info.type after
+ * // Tree.create(...) to choose the Ecrecover or Setter ratifier flow.
+ * console.log(info.type, info.ratifier);
  * ```
  */
 export async function fetchRatifierInfo(
@@ -34,7 +45,11 @@ export async function fetchRatifierInfo(
   const ecrecoverRatifier = getChainAddress(chainId, "ecrecoverRatifier");
   const setterRatifier = getChainAddress(chainId, "setterRatifier");
   const bytecode = await getBytecode(client, {
-    ...bytecodeCallParameters(params),
+    ...(params.blockNumber != null
+      ? { blockNumber: params.blockNumber }
+      : params.blockTag != null
+        ? { blockTag: params.blockTag }
+        : {}),
     address: params.maker,
   });
