+++ packages/midnight-sdk/src/signatures/SetterRatifierUtils.test.ts
@@ -1,0 +1,79 @@
+import { decodeFunctionData } from "viem";
+import { describe, expect, test } from "vitest";
+import { addresses, baseOffer } from "../__test__/fixtures.js";
+import { setterRatifierAbi } from "../abis.js";
+import { SetterRatifierUtils } from "./SetterRatifierUtils.js";
+import { Tree } from "./Tree.js";
+import { TreeUtils } from "./TreeUtils.js";
+
+const root =
+  "0x3333333333333333333333333333333333333333333333333333333333333333" as const;
+const proofNode =
+  "0x4444444444444444444444444444444444444444444444444444444444444444" as const;
+
+describe("SetterRatifierUtils.ratify", () => {
+  test("default", () => {
+    const offer = baseOffer({ maxAssets: 0n });
+    const tree = Tree.create(offer);
+
+    const items = SetterRatifierUtils.ratify({ tree });
+    const decoded = SetterRatifierUtils.decodeRatifierData(
+      items[0]!.ratifierData,
+    );
+
+    expect(items).toHaveLength(1);
+    expect(items[0]!.offer).toBe(offer);
+    expect(
+      TreeUtils.verifyProof({
+        offer,
+        group: items[0]!.group,
+        root: decoded.root,
+        leafIndex: decoded.leafIndex,
+        proof: decoded.proof,
+      }),
+    ).toBe(true);
+  });
+});
+
+describe("SetterRatifierUtils.encodeRatifierData", () => {
+  test("behavior: decode round trip", () => {
+    const data = SetterRatifierUtils.encodeRatifierData({
+      root,
+      leafIndex: 3n,
+      proof: [proofNode],
+    });
+
+    expect(SetterRatifierUtils.decodeRatifierData(data)).toEqual({
+      root,
+      leafIndex: 3n,
+      proof: [proofNode],
+    });
+  });
+});
+
+describe("SetterRatifierUtils.buildRootApprovalCall", () => {
+  test("default", () => {
+    const call = SetterRatifierUtils.buildRootApprovalCall({
+      setterRatifier: addresses.setterRatifier,
+      maker: addresses.maker,
+      root: "0x0000000000000000000000000000000000000000000000000000000000000000",
+    });
+    const decoded = decodeFunctionData({
+      abi: setterRatifierAbi,
+      data: call.data,
+    });
+
+    expect(decoded.functionName).toBe("setIsRootRatified");
+    expect(decoded.args[2]).toBe(true);
+  });
+
+  test("error: invalid bytes32 root", () => {
+    expect(() =>
+      SetterRatifierUtils.buildRootApprovalCall({
+        setterRatifier: addresses.setterRatifier,
+        maker: addresses.maker,
+        root: "0x1234",
+      }),
+    ).toThrow();
+  });
+});
