+++ packages/midnight-sdk/src/signatures/index.ts
@@ -1,5 +1,7 @@
-export * from "./OfferTreeUtils.js";
+export * from "./EcrecoverRatifierUtils.js";
+export * from "./GroupUtils.js";
 export * from "./Payload.js";
 export * from "./RatifierUtils.js";
+export * from "./SetterRatifierUtils.js";
 export * from "./Tree.js";
 export * from "./TreeUtils.js";
